requirements.txt
beautifulsoup4
python-dateutil
PyYAML
tqdm
requests
